+++ gsa/src/gmp/models/__tests__/filter.js
@@ -1117,7 +1117,7 @@ describe('Filter merge extra keywords', () => {
     const filter1 = Filter.fromString('abc=1');
     const filter2 = Filter.fromString(
       'apply_overrides=1 overrides=1 ' +
-        'autofp=1 delta_states=1 first=1 levels=hml min_qod=70 notes=1 ' +
+        'delta_states=1 first=1 levels=hml min_qod=70 notes=1 ' +
         'result_hosts_only=1 rows=10 sort=name timezone=CET',
     );
 
@@ -1126,7 +1126,6 @@ describe('Filter merge extra keywords', () => {
     expect(filter3.get('abc')).toBe('1');
     expect(filter3.get('apply_overrides')).toBe(1);
     expect(filter3.get('overrides')).toBe(1);
-    expect(filter3.get('autofp')).toBe(1);
     expect(filter3.get('delta_states')).toBe('1');
     expect(filter3.get('first')).toBe(1);
     expect(filter3.get('levels')).toBe('hml');
@@ -1139,10 +1138,8 @@ describe('Filter merge extra keywords', () => {
   });
 
   test('should merge new keywords', () => {
-    const filter1 = Filter.fromString('abc=1 autofp=0 trend=more');
-    const filter2 = Filter.fromString(
-      'autofp=1 delta_states=1 severity>3 sort=name',
-    );
+    const filter1 = Filter.fromString('abc=1 trend=more');
+    const filter2 = Filter.fromString('delta_states=1 severity>3 sort=name');
 
     const filter3 = filter2.mergeKeywords(filter1);
 
@@ -1150,7 +1147,6 @@ describe('Filter merge extra keywords', () => {
     expect(filter3.get('delta_states')).toBe('1');
     expect(filter3.get('sort')).toBe('name');
     expect(filter3.get('severity')).toBe('3');
-    expect(filter3.get('autofp')).toBe(1);
     expect(filter3.get('trend')).toBe('more');
   });
 
@@ -1223,7 +1219,7 @@ describe('Filter merge extra keywords', () => {
     filter1.id = 'f1';
     const filter2 = Filter.fromString(
       'apply_overrides=1 overrides=1 ' +
-        'autofp=1 delta_states=1 first=1 levels=hml min_qod=70 notes=1 ' +
+        'delta_states=1 first=1 levels=hml min_qod=70 notes=1 ' +
         'result_hosts_only=1 rows=10 sort=name timezone=CET',
     );
     filter2.id = 'f2';