+++ gsa/src/web/pages/extras/__tests__/cvsscalculatorpage.js
@@ -17,31 +17,10 @@
  */
 
 import React from 'react';
-import {act} from 'react-dom/test-utils';
 
-import {fireEvent, rendererWith, waitForElement} from 'web/utils/testing';
+import {fireEvent, rendererWith, waitFor, wait} from 'web/utils/testing';
 
 import CvssCalculator from 'web/pages/extras/cvsscalculatorpage';
-
-/* eslint-disable no-console */
-
-// this is just a little hack to silence a warning that we'll get until we
-// upgrade to 16.9: https://github.com/facebook/react/pull/14853
-const originalError = console.error;
-beforeAll(() => {
-  console.error = (...args) => {
-    if (/Warning.*not wrapped in act/.test(args[0])) {
-      return;
-    }
-    originalError.call(console, ...args);
-  };
-});
-
-afterAll(() => {
-  console.error = originalError;
-});
-
-/* eslint-enable no-console */
 
 const calculateScoreFromVector = jest.fn().mockReturnValue(
   Promise.resolve({
@@ -80,20 +59,34 @@ describe('CvssCalculator page tests', () => {
 
     const input = getAllByTestId('select-selected-value');
 
-    waitForElement(() => element.querySelectorAll('input'));
+    waitFor(() => element.querySelectorAll('input'));
     const vector = element.querySelectorAll('input');
 
+    /* CVSSv2 input */
     expect(input[0]).toHaveTextContent('Local');
     expect(input[1]).toHaveTextContent('Low');
     expect(input[2]).toHaveTextContent('None');
     expect(input[3]).toHaveTextContent('None');
     expect(input[4]).toHaveTextContent('None');
     expect(input[5]).toHaveTextContent('None');
+    /* CVSSv3 input */
+    expect(input[6]).toHaveTextContent('Network');
+    expect(input[7]).toHaveTextContent('Low');
+    expect(input[8]).toHaveTextContent('None');
+    expect(input[9]).toHaveTextContent('None');
+    expect(input[10]).toHaveTextContent('Unchanged');
+    expect(input[11]).toHaveTextContent('None');
+    expect(input[12]).toHaveTextContent('None');
+    expect(input[13]).toHaveTextContent('None');
 
     expect(vector[0]).toHaveAttribute('value', 'AV:L/AC:L/Au:N/C:N/I:N/A:N');
+    expect(vector[1]).toHaveAttribute(
+      'value',
+      'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    );
   });
 
-  test('Should render userVector from url', () => {
+  test('Should render userVector from url', async () => {
     const {render} = rendererWith({
       gmp,
       store: true,
@@ -103,19 +96,33 @@ describe('CvssCalculator page tests', () => {
       <CvssCalculator location={location} />,
     );
 
-    const input = getAllByTestId('select-selected-value');
-    waitForElement(() => element.querySelectorAll('input'));
+    await wait();
 
+    const input = getAllByTestId('select-selected-value');
     const vector = element.querySelectorAll('input');
 
+    /* CVSSv2 input */
     expect(input[0]).toHaveTextContent('Network');
     expect(input[1]).toHaveTextContent('Low');
     expect(input[2]).toHaveTextContent('None');
     expect(input[3]).toHaveTextContent('Partial');
     expect(input[4]).toHaveTextContent('Partial');
     expect(input[5]).toHaveTextContent('Partial');
+    /* CVSSv3 input */
+    expect(input[6]).toHaveTextContent('Network');
+    expect(input[7]).toHaveTextContent('Low');
+    expect(input[8]).toHaveTextContent('None');
+    expect(input[9]).toHaveTextContent('None');
+    expect(input[10]).toHaveTextContent('Unchanged');
+    expect(input[11]).toHaveTextContent('None');
+    expect(input[12]).toHaveTextContent('None');
+    expect(input[13]).toHaveTextContent('None');
 
     expect(vector[0]).toHaveAttribute('value', 'AV:N/AC:L/Au:N/C:P/I:P/A:P');
+    expect(vector[1]).toHaveAttribute(
+      'value',
+      'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    );
   });
 
   test('Changing userVector should change displayed select values', async () => {
@@ -128,25 +135,44 @@ describe('CvssCalculator page tests', () => {
       <CvssCalculator location={location} />,
     );
 
+    await wait();
+
     const vector = element.querySelectorAll('input');
 
-    act(() => {
-      fireEvent.change(vector[0], {
-        target: {value: 'AV:N/AC:L/Au:N/C:N/I:P/A:P'},
-      });
+    fireEvent.change(vector[0], {
+      target: {value: 'AV:N/AC:L/Au:N/C:N/I:P/A:P'},
     });
 
-    waitForElement(() => getAllByTestId('select-selected-value'));
+    fireEvent.change(vector[1], {
+      target: {value: 'CVSS:3.1/AV:P/AC:H/PR:L/UI:R/S:C/C:H/I:H/A:H'},
+    });
+
+    await wait();
+
     const input = getAllByTestId('select-selected-value');
 
+    /* CVSSv2 input */
     expect(input[0]).toHaveTextContent('Network');
     expect(input[1]).toHaveTextContent('Low');
     expect(input[2]).toHaveTextContent('None');
     expect(input[3]).toHaveTextContent('None');
     expect(input[4]).toHaveTextContent('Partial');
     expect(input[5]).toHaveTextContent('Partial');
+    /* CVSSv3 input */
+    expect(input[6]).toHaveTextContent('Physical');
+    expect(input[7]).toHaveTextContent('High');
+    expect(input[8]).toHaveTextContent('Low');
+    expect(input[9]).toHaveTextContent('Required');
+    expect(input[10]).toHaveTextContent('Changed');
+    expect(input[11]).toHaveTextContent('High');
+    expect(input[12]).toHaveTextContent('High');
+    expect(input[13]).toHaveTextContent('High');
 
     expect(vector[0]).toHaveAttribute('value', 'AV:N/AC:L/Au:N/C:N/I:P/A:P');
+    expect(vector[1]).toHaveAttribute(
+      'value',
+      'CVSS:3.1/AV:P/AC:H/PR:L/UI:R/S:C/C:H/I:H/A:H',
+    );
   });
 
   test('Changing displayed select values should change userVector', async () => {
@@ -159,19 +185,34 @@ describe('CvssCalculator page tests', () => {
       <CvssCalculator location={location} />,
     );
 
+    await wait();
+
     const vector = element.querySelectorAll('input');
 
-    waitForElement(() => getAllByTestId('select-selected-value'));
     const input = getAllByTestId('select-selected-value');
 
+    /* CVSSv2 input */
     expect(input[0]).toHaveTextContent('Network');
     expect(input[1]).toHaveTextContent('Low');
     expect(input[2]).toHaveTextContent('None');
     expect(input[3]).toHaveTextContent('Partial');
     expect(input[4]).toHaveTextContent('Partial');
     expect(input[5]).toHaveTextContent('Partial');
+    /* CVSSv3 input */
+    expect(input[6]).toHaveTextContent('Network');
+    expect(input[7]).toHaveTextContent('Low');
+    expect(input[8]).toHaveTextContent('None');
+    expect(input[9]).toHaveTextContent('None');
+    expect(input[10]).toHaveTextContent('Unchanged');
+    expect(input[11]).toHaveTextContent('None');
+    expect(input[12]).toHaveTextContent('None');
+    expect(input[13]).toHaveTextContent('None');
 
     expect(vector[0]).toHaveAttribute('value', 'AV:N/AC:L/Au:N/C:P/I:P/A:P');
+    expect(vector[1]).toHaveAttribute(
+      'value',
+      'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    );
 
     const selectFields = getAllByTestId('select-open-button');
 
@@ -181,6 +222,18 @@ describe('CvssCalculator page tests', () => {
 
     fireEvent.click(selectItems[0]);
 
+    const selectFields2 = getAllByTestId('select-open-button');
+
+    fireEvent.click(selectFields2[7]);
+
+    const selectItems2 = getAllByTestId('select-item');
+
+    fireEvent.click(selectItems2[1]);
+
     expect(vector[0]).toHaveAttribute('value', 'AV:L/AC:L/Au:N/C:P/I:P/A:P');
+    expect(vector[1]).toHaveAttribute(
+      'value',
+      'CVSS:3.1/AV:N/AC:H/PR:N/UI:N/S:U/C:N/I:N/A:N',
+    );
   });
 });