+++ gsa/src/web/pages/scanners/dialog.js
@@ -54,36 +54,19 @@ import {getTimezone} from 'web/store/usersettings/selectors';
 
 import {
   OSP_SCANNER_TYPE,
-  GMP_SCANNER_TYPE,
   GREENBONE_SENSOR_SCANNER_TYPE,
   scannerTypeName,
 } from 'gmp/models/scanner';
 
-import {
-  CLIENT_CERTIFICATE_CREDENTIAL_TYPE,
-  USERNAME_PASSWORD_CREDENTIAL_TYPE,
-} from 'gmp/models/credential';
+import {CLIENT_CERTIFICATE_CREDENTIAL_TYPE} from 'gmp/models/credential';
 
 const AVAILABLE_SCANNER_TYPES = [
   OSP_SCANNER_TYPE,
-  GMP_SCANNER_TYPE,
   GREENBONE_SENSOR_SCANNER_TYPE,
 ];
 
 const client_cert_credentials_filter = credential => {
   return credential.credential_type === CLIENT_CERTIFICATE_CREDENTIAL_TYPE;
-};
-
-const username_password_credentials_filter = credential => {
-  return credential.credential_type === USERNAME_PASSWORD_CREDENTIAL_TYPE;
-};
-
-const filter_credentials = (credentials, type) => {
-  const cred_filter =
-    type === GMP_SCANNER_TYPE
-      ? username_password_credentials_filter
-      : client_cert_credentials_filter;
-  return filter(credentials, cred_filter);
 };
 
 const render_certificate_info = (info, tz) => {
@@ -148,7 +131,10 @@ const ScannerDialog = ({
   const handleTypeChange = (value, name) => {
     if (onScannerTypeChange) {
       value = parseInt(value);
-      const scan_credentials = filter_credentials(credentials, value);
+      const scan_credentials = filter(
+        credentials,
+        client_cert_credentials_filter,
+      );
 
       onScannerTypeChange(value, name);
       onScannerTypeChange(
@@ -173,13 +159,9 @@ const ScannerDialog = ({
   const {gmp} = props;
 
   if (gmp.settings.enableGreenboneSensor) {
-    SCANNER_TYPES = [
-      GMP_SCANNER_TYPE,
-      OSP_SCANNER_TYPE,
-      GREENBONE_SENSOR_SCANNER_TYPE,
-    ];
+    SCANNER_TYPES = [OSP_SCANNER_TYPE, GREENBONE_SENSOR_SCANNER_TYPE];
   } else {
-    SCANNER_TYPES = [GMP_SCANNER_TYPE, OSP_SCANNER_TYPE];
+    SCANNER_TYPES = [OSP_SCANNER_TYPE];
   }
 
   let {credential_id} = props;
@@ -189,7 +171,10 @@ const ScannerDialog = ({
     value: scannerType,
   }));
 
-  const scanner_credentials = filter_credentials(credentials, type);
+  const scanner_credentials = filter(
+    credentials,
+    client_cert_credentials_filter,
+  );
   const is_edit = isDefined(scanner);
   const isInUse = isDefined(scanner) && scanner.isInUse();
   const show_cred_info =