+++ gsad/src/gsad.c
@@ -477,8 +477,6 @@ init_validator ()
   gvm_validator_add (validator, "authdn", "^.{0,200}%s.{0,200}$");
   gvm_validator_add (validator, "auto_delete", "^(no|keep)$");
   gvm_validator_add (validator, "auto_delete_data", "^.*$");
-  gvm_validator_add (validator, "autofp", "^(0|1|2)$");
-  gvm_validator_add (validator, "autofp_value", "^(1|2)$");
   gvm_validator_add (validator, "boolean", "^(0|1)$");
   gvm_validator_add (validator, "bulk_selected:name", "^.*$");
   gvm_validator_add (validator, "bulk_selected:value", "(?s)^.*$");
@@ -709,7 +707,6 @@ init_validator ()
   gvm_validator_add (validator, "schedule_id", "^[a-z0-9\\-]+$");
   gvm_validator_add (validator, "severity",
                      "^(-1(\\.0)?|[0-9](\\.[0-9])?|10(\\.0)?)$");
-  gvm_validator_add (validator, "severity_class", "^(nist|pci\\-dss)$");
   gvm_validator_add (validator, "severity_optional",
                      "^(-1(\\.0)?|[0-9](\\.[0-9])?|10(\\.0)?)?$");
   gvm_validator_add (validator, "source_iface", "^(.*){1,16}$");
@@ -3283,7 +3280,7 @@ main (int argc, char **argv)
     {
       if (termination_signal)
         {
-          g_debug ("Received %s signal.\n", sys_siglist[termination_signal]);
+          g_debug ("Received %s signal.\n", strsignal (termination_signal));
           gsad_cleanup ();
           /* Raise signal again, to exit with the correct return value. */
           signal (termination_signal, SIG_DFL);